Remove incorrect forced_unique factory.
The unique_keys constructor argment for Union needs to be the set of
naturally unique keys, since that's something it can't compute from
its members (but something higher-level code can know).  That means
we can't use it as a relation that forces uniqueness via SELECT
DISTINCT or SQL UNION (instead of UNION ALL).  I'll bring back a
dedicated ForcedUnique relation in the future.